chore: remove pkg/migration
It is no longer needed since it was used for monolithic -> family
providers many months ago and is no longer valuable.

It's a good time to do this now as it has code that is not compatible
with Crossplane v2, e.g. it migrates classic patch & transform
compositions, which are removed in v2.

Signed-off-by: Jared Watts <jbw976@gmail.com>